Bugs fixed

+++ lms/src/main/java/com/project/lms/Dto/StaffResponse.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @AllArgsConstructor
 public class StaffResponse {
 
-    private int staffId;
+    private Long staffId;
     private String name;
 }
